+++ src/components/horizontal-bar-chart/horizontal-bar-chart.ts
@@ -8,9 +8,9 @@ import { clearChart, generateChart } from '.'
 
 export default class HorizontalBarChartComponent extends Vue {
   @Prop({ default: [] }) chartData!: ChartModels.GroupedBarChartItem[]
-  @Prop({ default: 'chart' }) domId!: string
-  @Prop({ default: '' }) chartTitle!: string
-  @Prop({ default: '' }) chartExportName!: string
+  @Prop() domId!: string
+  @Prop() chartTitle!: string
+  @Prop() chartExportName!: string
 
   get hasData (): boolean {
     return this.chartData.length > 0
@@ -29,7 +29,7 @@ export default class HorizontalBarChartComponent extends Vue {
 
   renderChart (): void {
     const id = this.domId
-    const screenWidth = (document.getElementById('horizontal-bar-chart-component')?.clientWidth ?? 0)
+    const screenWidth = (document.getElementById(`wrapper-${this.domId}`)?.clientWidth ?? 0)
     const config = {
       width: screenWidth,
       margins: { top: 20, right: 20, bottom: 30, left: 80 },
@@ -52,7 +52,6 @@ export default class HorizontalBarChartComponent extends Vue {
 
     // TODO: 109 add legend if needed
     // TODO: 107 function to compute shortname of dataset to add to legend
-
     setTimeout(() => {
       void exportChartWithElement(chart, this.chartExportName)
     }, 200)