+++ src/components/navbar/mobile-menu-toggle-button/mobile-menu-toggle-button.ts
@@ -1,6 +1,6 @@
 import { Vue } from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
 
-export default class MobileMenuToggleButton extends Vue {
+export default class MobileMenuToggleButtonComponent extends Vue {
   @Prop({ default: false }) readonly isToggled!: boolean
 }