+++ src/models/Chart.ts
@@ -1,7 +1,3 @@
-import { Dayjs } from 'dayjs'
-
-import { Site } from './Site'
-
 export interface ChartSVGElement {
   svg: SVGSVGElement
   width: number
@@ -17,28 +13,4 @@ export interface BarChartItem {
 export interface GroupedBarChartItem {
   group: string
   series: BarChartItem[]
-}
-
-export interface MapDataSet {
-  startDate: Dayjs
-  endDate: Dayjs
-  sites: Site[]
-  color: string
-  data: MapSiteData[]
-}
-
-export interface MapSiteData {
-  siteName: string
-  longitude: number
-  latitude: number
-  distinctSpecies: { [taxonName: string]: number }
-}
-
-export interface TableData {
-  speciesSlug: string
-  speciesId: number
-  speciesName: string
-  className: string
-  data: boolean[]
-  total: number
 }