+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * **features:** Detected species by time (#20)
 * **features:** Export csv support multiple datasets(#106)
 * **features:** Each species row in table should be clickable and link to ap page (#110)
+* **features:** Export line chart as PNG (#85)
 * **features:** Function to compute shortname to describe dataset and update export filename(#107, #108)
 
 * **refactor:** Refactor species richness table display (#112)